http!: Change Http trait method to just `send`
This makes for a cleaner UX by removing the need for
both `get` and `post`.

To make it work, the `Http::send` method also accepts a
`impl Into<Self::Body>` as the body parameter. We introduce
an additional bound `Body: From<Vec<u8>>`. To send a
request with no body one can simply pass an empty Vec.

Added new type `HttpMethod` for specifying the HTTP method
to be used in the request (e.g. GET, POST).

Refactor the request client config builder to be more
idiomatic.